+++ docs/changes.rst
@@ -77,8 +77,8 @@ Released: not yet
   commands have been grouped to be more easily identifiable. This required
   adding the "click-option-group" Python package to the dependencies.
 
-* Increased minimum zhmcclient version to 1.12.1 to pick up fixes and
-  functionality. (issue #510)
+* Increased minimum zhmcclient version to 1.13.3 to pick up fixes and
+  functionality. (issues #510, #528)
 
 * Tests: Added an environment variable TESTLOG to enable logging for end2end
   tests. (issue #414)
@@ -117,6 +117,12 @@ Released: not yet
 
 * Added support for user patterns with a new 'zhmc userpattern' command group.
   (issue #550)
+
+* Added support for installation of single firmware updates on the SE with a
+  new 'zhmc cpc install-firmware' command. (issue #528)
+
+* Added support for deletion of uninstalled firmware updates from the SE with a
+  new 'zhmc cpc delete-uninstalled-firmware' command. (issue #528)
 
 **Cleanup:**